Refactor internal datapack.
The previous folder structures grouped power JSONs by which origin used them, this has been changed so that they are instead grouped by context.
Powers that rely on others (such as the Death's Vice or Floral Mimicry powers) and powers that are very contextual to their origin (such as the Jelly's Sculk Spread) remain in their existing folders. However more generic powers such as Pescatarian and Double Jump have been moved to a common folder to indicate that they can safely be reused in other origins.
For documentation purposes, only the powers under the common folder will be explicitly listed and described in "base contents" in the GitHub repo's wiki.