Bugfix the double jump common power.
In testing, I discovered a likely bug in Apoli. Landing packets are not sent to the server for entities with any form of slow fall (power or status effect); this created potential issues with the Jelly Sculk and Faerie Moth. The double jump power has been adjusted to create its own landing identifier instead of using the in-built one.