Update to 1.20
This will not build due on JitPack or GitHub due to dependency issues, it's a miracle that it builds locally at all.
This commit is largely just adjusting to new variable visibility in 1.20. The vibration power has been internally overhauled since Mojang changed the vibration system again - but the API remains the same except for a couple new additional fields.